install local autokeras + update tensorflow and keras-tuner version
This commit halves the size of the autokeras Docker image. TensorFlow is updated to 2.3.0, and keras-tuner is updated to 1.0.2rc1. In addition, the local copy of autokeras is used. This is helpful if changes are made to autokeras and a developer would like to test the changes in a Docker image, building the Docker image will use the local copy of the code. Last, the cache is not used in `pip install` to reduce the size of the built image.